refactor: split helpers and middleware
