+++ web/app.py
@@ -31,12 +31,13 @@ from aiohttp.web_exceptions import HTTPForbidden
 from aiohttp_session import get_session
 from jinja2 import pass_context
 from aiohttp_jinja2 import static_root_key
-from aiolimiter import AsyncLimiter
-from collections import defaultdict
 import io, qrcode, pyotp
 from PIL import Image
 import subprocess
 from pdf2image import convert_from_path
+from web.utils import render as _render, sign_token as _sign_token, verify_token as _verify_token, issue_csrf, notify_shared_upload
+from web.tasks import generate_hls as _generate_hls
+from web.middleware import csrf_protect_mw, auth_mw, csp_mw, rl_mw
 
 from bot.db import init_db  # スキーマ初期化用
 
@@ -85,258 +86,6 @@ MOBILE_TEMPLATES = {
 }
 
 
-def _is_mobile(user_agent: str) -> bool:
-    if not user_agent:
-        return False
-    pattern = (
-        r"iPhone|Android.*Mobile|Windows Phone|iPod|BlackBerry|Opera Mini|IEMobile"
-    )
-    return re.search(pattern, user_agent, re.I) is not None
-
-
-def _render(req: web.Request, tpl: str, ctx: Dict[str, object]):
-    ctx.setdefault("user_id", req.get("user_id"))
-    ua = req.headers.get("User-Agent", "")
-    if _is_mobile(ua):
-        tpl = MOBILE_TEMPLATES.get(tpl, tpl)
-    return aiohttp_jinja2.render_template(tpl, req, ctx)
-
-
-def _sign_token(fid: str, exp: int) -> str:
-    """Generate a signed token for file downloads."""
-    msg = f"{fid}:{exp}".encode()
-    sig = hmac.new(FILE_HMAC_SECRET, msg, hashlib.sha256).digest()
-    return base64.urlsafe_b64encode(msg + b":" + sig).decode()
-
-
-def _verify_token(tok: str) -> Optional[str]:
-    """Validate a token and return the file id if it is valid."""
-    try:
-        raw = base64.urlsafe_b64decode(tok.encode())
-        fid, exp_raw, sig = raw.split(b":", 2)
-
-        # 無期限トークンは exp==0
-        exp_val = int(exp_raw)
-        if exp_val != 0 and exp_val < int(datetime.now(timezone.utc).timestamp()):
-            return None
-        valid = hmac.compare_digest(
-            sig,
-            hmac.new(
-                FILE_HMAC_SECRET,
-                f"{fid.decode()}:{exp_raw.decode()}".encode(),
-                hashlib.sha256,
-            ).digest(),
-        )
-        return fid.decode() if valid else None
-    except Exception:
-        return None
-
-
-async def issue_csrf(request: web.Request) -> str:
-    """Return a CSRF token stored in the session, creating one if needed."""
-    session = await aiohttp_session.get_session(request)
-    if "csrf_token" not in session:
-        session["csrf_token"] = secrets.token_urlsafe(16)
-    return session["csrf_token"]
-
-
-async def _send_shared_webhook(db: Database, folder_id: int, message: str) -> None:
-    """指定フォルダの Webhook にメッセージを送信"""
-    rec = await db.get_shared_folder(int(folder_id))
-    url = rec["webhook_url"] if rec else None
-    if not url:
-        return
-    async with aiohttp.ClientSession() as session:
-        await session.post(url, json={"content": message})
-
-
-async def notify_shared_upload(
-    db: Database, folder_id: int, discord_id: int, file_name: str
-) -> None:
-    """共有フォルダへのアップロードをWebhookで通知"""
-    mention = f"<@{discord_id}>"
-    message = f"\N{INBOX TRAY} {mention} が `{file_name}` をアップロードしました。"
-    await _send_shared_webhook(db, folder_id, message)
-
-
-# ─────────────── Background Processing ───────────────
-def _generate_preview_and_tags(path: Path, fid: str, file_name: str) -> str:
-    """Create preview image and return tags."""
-    mime, _ = mimetypes.guess_type(file_name)
-    preview_path = PREVIEW_DIR / f"{fid}.jpg"
-    try:
-        if mime and mime.startswith("image"):
-            img = Image.open(path)
-            img.thumbnail((320, 320))
-            img.convert("RGB").save(preview_path, "JPEG")
-        elif mime and mime.startswith("video"):
-            subprocess.run(
-                [
-                    "ffmpeg",
-                    "-y",
-                    "-i",
-                    str(path),
-                    "-ss",
-                    "00:00:01",
-                    "-vframes",
-                    "1",
-                    str(preview_path),
-                ],
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-        elif mime == "application/pdf":
-            pages = convert_from_path(str(path), first_page=1, last_page=1)
-            if pages:
-                img = pages[0]
-                img.thumbnail((320, 320))
-                img.save(preview_path, "JPEG")
-        elif mime and mime.startswith("application/vnd"):
-            tmp_pdf = path.with_suffix(".pdf")
-            subprocess.run(
-                [
-                    "libreoffice",
-                    "--headless",
-                    "--convert-to",
-                    "pdf",
-                    str(path),
-                    "--outdir",
-                    str(path.parent),
-                ],
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-            if tmp_pdf.exists():
-                pages = convert_from_path(str(tmp_pdf), first_page=1, last_page=1)
-                if pages:
-                    img = pages[0]
-                    img.thumbnail((320, 320))
-                    img.save(preview_path, "JPEG")
-                tmp_pdf.unlink(missing_ok=True)
-        else:
-            preview_path = None
-    except Exception as e:
-        log.warning("preview generation failed: %s", e)
-        if preview_path and preview_path.exists():
-            preview_path.unlink(missing_ok=True)
-    from bot.auto_tag import generate_tags
-
-    return generate_tags(path, file_name)
-
-
-async def _generate_hls(path: Path, fid: str) -> None:
-    """Create HLS streams for the given video."""
-    variants = [
-        ("360p", 640, 360, 800_000),
-        ("720p", 1280, 720, 2_400_000),
-    ]
-    out_dir = HLS_DIR / fid
-    out_dir.mkdir(parents=True, exist_ok=True)
-    for name, w, h, br in variants:
-        out = out_dir / f"{name}.m3u8"
-        proc = await asyncio.create_subprocess_exec(
-            "ffmpeg",
-            "-y",
-            "-i",
-            str(path),
-            "-vf",
-            f"scale=w={w}:h={h}",
-            "-c:v",
-            "libx264",
-            "-c:a",
-            "aac",
-            "-b:v",
-            str(br),
-            "-hls_time",
-            "4",
-            "-hls_playlist_type",
-            "vod",
-            str(out),
-            stdout=asyncio.subprocess.DEVNULL,
-            stderr=asyncio.subprocess.DEVNULL,
-        )
-        await proc.wait()
-    master = out_dir / "master.m3u8"
-    with master.open("w") as f:
-        f.write("#EXTM3U\n#EXT-X-VERSION:3\n")
-        for name, w, h, br in variants:
-            f.write(f"#EXT-X-STREAM-INF:BANDWIDTH={br},RESOLUTION={w}x{h}\n")
-            f.write(f"{name}.m3u8\n")
-
-
-async def _task_worker(app: web.Application):
-    """Worker coroutine processing preview and tagging tasks."""
-    queue: asyncio.Queue = app["task_queue"]
-    while True:
-        job = await queue.get()
-        try:
-            tags = await asyncio.to_thread(
-                _generate_preview_and_tags, job["path"], job["fid"], job["file_name"]
-            )
-            if job.get("shared"):
-                await app["db"].update_shared_tags(job["fid"], tags)
-            else:
-                await app["db"].update_tags(job["fid"], tags)
-        except Exception as e:
-            log.exception("Background task failed: %s", e)
-        finally:
-            queue.task_done()
-
-
-# ─────────────── Middleware ───────────────
-@web.middleware
-async def csrf_protect_mw(request: web.Request, handler):
-    if request.method in ("POST", "PUT", "PATCH", "DELETE"):
-        session = await aiohttp_session.get_session(request)
-
-        # ヘッダー優先。なければフォームから取得。
-        token = request.headers.get("X-CSRF-Token")
-        if token is None:
-            form = await request.post()
-            token = form.get("csrf_token")
-
-        if token != session.get("csrf_token"):
-            if "application/json" in request.headers.get("Accept", ""):
-                return web.json_response(
-                    {"success": False, "error": "invalid csrf"}, status=403
-                )
-            raise HTTPForbidden(text="Invalid CSRF token")
-    return await handler(request)
-
-
-@web.middleware
-async def auth_mw(request: web.Request, handler):
-    sess = await aiohttp_session.get_session(request)
-    request["user_id"] = sess.get("user_id")
-    return await handler(request)
-
-
-CSP_POLICY = (
-    "default-src 'self'; "
-    "script-src 'self' https://cdn.jsdelivr.net https://cdnjs.cloudflare.com 'unsafe-inline'; "
-    "style-src 'self' https://cdn.jsdelivr.net https://cdnjs.cloudflare.com 'unsafe-inline'; "
-    "img-src 'self' data:; "
-    "font-src 'self' https://cdn.jsdelivr.net https://cdnjs.cloudflare.com; "
-    "connect-src 'self' https://cdn.jsdelivr.net https://cdnjs.cloudflare.com;"
-)
-
-
-@web.middleware
-async def csp_mw(request: web.Request, handler):
-    resp = await handler(request)
-    resp.headers["Content-Security-Policy"] = CSP_POLICY
-    return resp
-
-
-limiters = defaultdict(lambda: AsyncLimiter(30, 60))  # 60 秒あたり 30 リクエスト / IP
-
-
-@web.middleware
-async def rl_mw(req, handler):
-    ip = req.remote
-    limiter = limiters[ip]
-    async with limiter:
-        return await handler(req)
 
 
 # ─────────────── APP Factory ───────────────
